# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Group AI Hub API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type AgentBlock = (typeof AgentBlock)[keyof typeof AgentBlock];

export const AgentBlock = {
  estrategia: "estrategia",
  prospeccao: "prospeccao",
  marketing: "marketing",
  gestao: "gestao",
} as const;

export interface Agent {
  id: string;
  name: string;
  slug: string;
  description: string;
  block: AgentBlock;
  blockLabel: string;
  icon: string;
  systemPrompt: string;
  suggestedPrompts: string[];
  priority: number;
  color: string;
  designStudio?: boolean;
}

export interface Conversation {
  id: string;
  agentId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export type MessageMetadata = { [key: string]: unknown };

export interface Message {
  id: string;
  conversationId: string;
  role: MessageRole;
  content: string;
  createdAt: string;
  metadata?: MessageMetadata;
}

export interface ConversationWithMessages {
  id: string;
  agentId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  model: string;
  provider: string;
  agentName: string;
  messages: Message[];
}

export interface CreateConversationRequest {
  agentId: string;
  title?: string;
  model?: string;
}

export interface RenameConversationRequest {
  title: string;
}

export interface RenameConversationResponse {
  id: string;
  agentId: string;
  title: string;
  updatedAt: string;
}

export interface SendMessageRequest {
  content: string;
  useKnowledgeBase?: boolean;
  customSystemPrompt?: string;
  model?: string;
  stream?: boolean;
}

export interface MessageResponse {
  userMessage: Message;
  assistantMessage: Message;
  autoTitle?: string | null;
  model: string;
  provider: string;
}

export type KnowledgeDocumentStatus =
  (typeof KnowledgeDocumentStatus)[keyof typeof KnowledgeDocumentStatus];

export const KnowledgeDocumentStatus = {
  pending: "pending",
  processing: "processing",
  processed: "processed",
  error: "error",
} as const;

export interface KnowledgeDocument {
  id: string;
  agentId: string;
  filename: string;
  fileType: string;
  fileSize: number;
  storageKey: string;
  status: KnowledgeDocumentStatus;
  hasContent?: boolean;
  processed?: boolean;
  jobId?: string | null;
  createdAt: string;
}

export interface KnowledgeUploadResponse {
  success: boolean;
  message: string;
  document: KnowledgeDocument;
}

export interface UploadUrlRequest {
  agentId: string;
  filename: string;
  fileType: string;
  fileSize: number;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  documentId: string;
  storageKey: string;
}

export interface GenerateImageRequest {
  prompt: string;
  style?: string;
  agentId?: string;
}

export interface GalleryImage {
  url: string;
  prompt: string;
  createdAt: string;
}

export interface GenerateImageResponse {
  imageUrl: string;
  prompt: string;
  gallery?: GalleryImage[];
}

export interface ImageGalleryResponse {
  images: GalleryImage[];
}

export interface CanvaLinkRequest {
  contentType: string;
  title?: string;
  description?: string;
}

export interface CanvaLinkResponse {
  url: string;
  contentType: string;
  label?: string;
}

export interface SearchKnowledgeRequest {
  query: string;
  agentId?: string;
  limit?: number;
}

export type SearchKnowledgeResponseResultsItem = {
  documentId: string;
  filename: string;
  content: string;
  score: number;
};

export interface SearchKnowledgeResponse {
  results: SearchKnowledgeResponseResultsItem[];
  query: string;
}

export interface IntegrationItem {
  id: string;
  name: string;
  description: string;
  envVar: string;
  configured: boolean;
  active: boolean;
  category: string;
}

export interface IntegrationSettingsResponse {
  integrations: IntegrationItem[];
  activeLLM?: string | null;
  ollamaModel: string;
  openrouterModel: string;
  activeProvider?: string;
  activeModel?: string | null;
}

export type ActiveProviderResponseSource =
  (typeof ActiveProviderResponseSource)[keyof typeof ActiveProviderResponseSource];

export const ActiveProviderResponseSource = {
  tenant: "tenant",
  legacy: "legacy",
  default: "default",
} as const;

export interface ActiveProviderResponse {
  provider: string;
  customUrl: string | null;
  model: string | null;
  source: ActiveProviderResponseSource;
}

export interface SetActiveProviderRequest {
  provider: string;
  customUrl?: string;
  model?: string;
}

export type ActiveProviderSaveResponse = ActiveProviderResponse & {
  success: boolean;
};

export interface TestActiveProviderRequest {
  provider?: string;
  customUrl?: string;
  model?: string;
}

export interface TestActiveProviderResponse {
  success: boolean;
  response?: string;
  provider?: string;
  model?: string;
  tokensUsed?: number;
  executionTimeMs?: number;
  error?: string;
}

export interface ModelOption {
  id: string;
  name: string;
  description: string;
}

export interface AvailableModelsResponse {
  models: ModelOption[];
  defaultModel: string;
  provider?: string | null;
}

export interface CustomKey {
  provider: string;
  createdAt: string;
  updatedAt?: string;
  keyLast4?: string | null;
}

export interface SetCustomKeyRequest {
  provider: string;
  key: string;
}

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Internal server error
 */
export type ServerErrorResponse = ErrorResponse;

export type ListAgents200 = {
  agents: Agent[];
};

export type ListConversationsParams = {
  agentId?: string;
};

export type ListConversations200 = {
  conversations: Conversation[];
};

export type ListKnowledgeDocumentsParams = {
  agentId?: string;
};

export type ListKnowledgeDocuments200 = {
  documents: KnowledgeDocument[];
};

export type UploadKnowledgeDocumentBody = {
  file: Blob;
  agentId: string;
};

export type GetCustomKeys200 = {
  keys: CustomKey[];
};