# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Tax Group AI Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /knowledge:
    get:
      operationId: listKnowledgeDocuments
      tags: [knowledge]
      summary: List knowledge base documents
      parameters:
        - name: agentId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/KnowledgeDocument"
                required: [documents]
  /knowledge/upload:
    post:
      operationId: uploadKnowledgeDocument
      tags: [knowledge]
      summary: Upload a knowledge document with text extraction
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                agentId:
                  type: string
              required: [file, agentId]
      responses:
        "202":
          description: Document accepted and queued for background processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeUploadResponse"
...
  /knowledge/{documentId}:
    delete:
      operationId: deleteKnowledgeDocument
      tags: [knowledge]
      summary: Delete a knowledge document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    KnowledgeDocument:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        filename:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        storageKey:
          type: string
        status:
          type: string
          enum: [pending, processing, processed, error]
        hasContent:
          type: boolean
        processed:
          type: boolean
        jobId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, agentId, filename, fileType, fileSize, storageKey, status, createdAt]
    KnowledgeUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        document:
          $ref: "#/components/schemas/KnowledgeDocument"
      required: [success, message, document]
...