# lib/api-zod/src/generated/types/knowledgeDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Group AI Hub API
 * OpenAPI spec version: 0.2.0
 */
import type { KnowledgeDocumentStatus } from "./knowledgeDocumentStatus";

export interface KnowledgeDocument {
  id: string;
  agentId: string;
  filename: string;
  fileType: string;
  fileSize: number;
  storageKey: string;
  status: KnowledgeDocumentStatus;
  hasContent?: boolean;
  processed?: boolean;
  jobId?: string | null;
  createdAt: Date;
}