# lib/api-zod/src/generated/types/knowledgeUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax Group AI Hub API
 * OpenAPI spec version: 0.2.0
 */
import type { KnowledgeDocument } from "./knowledgeDocument";

export interface KnowledgeUploadResponse {
  success: boolean;
  message: string;
  document: KnowledgeDocument;
}